Add files via upload
The assignment of the age variable relies on correct user input.
        The assignment of the age variable relies on correct user input. If the user enters characters instead of digits, a runtime error causes the program to stop immediately. Note that this will not cause a syntax error when you compile your program. To catch this and exit gracefully, add a try clause before the assignment of the age variable. Add a catch clause for a generic exception and let the user know they must enter a whole number only.  Then, set age to 0 and let your program continue, picking up on the validation loop.

Add a catch clause here for a generic exception and let the user know they must enter a whole number only.  Then, set age to 0 and let your program continue, returning the value of age to main() where it is then validated.

If anything other than a number is entered, the catch clause should set the age to 0. Once age is set - either what is entered by the user OR a 0 set by the catch clause, return it